BackendServer/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
backports.zoneinfo==0.2.1
cffi==1.15.0
channels==3.0.3
channels-redis==3.3.1
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.0.2
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mongoengine==0.23.1
msgpack==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==3.12.3
pyOpenSSL==22.0.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
zope.interface==5.4.0
